change lineplot to tsx and try adding axes; doesn't work yet
